Tests/LibWeb/Text/expected/Streams/ReadableStream-read-subview-chunk.txt

length=3 bytes=12,13,14
